Checkpoint - A Crew can be Selected
Holy crap. This has been a rather awful journey trying to learn
React, Redux, and this particular packing scheme. I understand the
value in starting from something like this... But seriously, HOLY
CRAP!

Okay, let's get professional or something. What we have so far is
stubbed in methods to retrieve fake data generated by the stuff in
/server/resources. This is loaded in the site/builder route.

... Crap... Getting professional didn't happen.

From there, it iterates over the current fake list of crews, then
displays them. When you click on one of the <li>s, you then get
the crew name selected to appear beneath. This is accomplished via
a redux action. This is the cool part, I think.

My problem at the moment is that it seems like I'm cruising
towards a monolithic reducer for the builder. This seems to be
perfectly counter to redux. I tried to decompose (is that the right
word? 'tis late at night after a long day reacting-reduxing) the
reducer into multiple files unsuccessfully.

There's a lot to be learned yet. A basic list is:
 - How do I further decompose reducers.
 - How can I reuse reducers, or even state, between /s (which I am
   non-sarcastically reading as "whacks"... Ooph).
 - How can I stick some basic JSON files to read in? I would much
   rather do that than hard-code crap into one massive reducer.
   Insert another Ooph here... Yikes.

OH! And stop being so darn lazy on your commits and branching!

USE DA CM, MANG!

... Also, where the heck is the testing of your component? One of
the most exciting parts of redux to me is how it should make
testing "business logic" so darn simple. C'mon duder...

Okay, gl;hf!

... definitely failed at professionalism... GJ...