Adding rap sheets!
Yay!

We did it, team!

... or, we did it, just me!

Rap sheets are now being added when they can. Looks good so far.

I think that Traits are going to need to be separated into their
own Component in order to move some of the string processing
knowledge that may be platform dependant somewhere else.

So this only works at the moment if you select Batman.

Next up is populating a few more models to do some additional crew
building logic.