Dynamic traits are now implemented
In other words, the text of a trait can now be modified during trait
loading. This allows for more diverse trait support for things like
Affinity, Primary Target, etc.

It does require some knowledge of how it is implemented
unfortunately. In other words, I've increased the tribal knowledge
required to improve the program.

On the plus side, I now have a tribe! Huzzah!

... even if it's just me...