Long overdue refactoring and unit testing
Unit tests!!!!

... Finally... Lazy developer...

Also, some pretty extensive refactoring.

To get unit test working (and honestly, I wasted a good 5 hours due
to the original, BAD javascript viewer that was being used for the
testing framework... Switch to Firefox/Chrome immediately in the
future!) and make them better tests, a lot of work has happened.

Too much to talk about.

But it's closer to working than it's ever been before...

There is a nasty bug on the de-select... Yeah, don't deselect for
now.