Refactoring of resources and better unit tests
Finally. The unit tests are paying off on the crewRulesEngine.

Man, non-destructive removing of characters from one array to
another is a complete and total pain. I just need to stop being
such a wimp and do it destructively. The code will be far simpler.

Anywho, the bigger change is that resources are now being loaded
into the resource loader directly instead of having a conversion
path somewhere in the middle. This is good. Less boring and
repetitive work is always good.

Pleased with how the logic is coming along.