First prototype print is a success
Not even april foolins... It's a success!

Unfortunately, the font is completely unreadable. So... Almost a
success?

Honestly, I need to compare the font size to Malifaux cards, I
think it may be larger...

Also, I believe that I can start to refine the spacing of the
whole thing to get a better result. Tuning the margins more
correctly will allow for better looking sheets.

Also, I want to determine how I can dynamically size the font so
that it's not quite so tiny for people that aren't Batman. Even
better would be to split it onto two sheets dynamically for those
that are Batman...

All in all, I'm thrilled with the current state of the prototype.
Less than a week's time has been spent on it, and I have my first
true test.

Making a note here: HUGE SUCCESS!!!@