Move selection into its own header
Kinda the last big functionality change before gutting the app and
cleaning up.

The next change is going to be a ton of deletes. gl;hf!

This change was mostly about moving crew selection to a select
box... which isn't as good looking as I hoped. I made the incorrect
assumption that there was some cool stying in bootstrap that would
make it look better. Oh well, the CSS will have to follow...

... because man, pixel pushing in CSS is painful and boring. We as
a programming community have to fix it.