Completed the Great Purge (tm)
Well, not actually trademarked.

But the purge has been completed. Much of the original, unnecessary
stuff from the original fork is no longer there.

It's good. Definitely an improvement.

It also motivated me to update the CSS a little bit. Things are
looking better now.

So much remains to do. I'll probably focus on refactoring the
character select to make it more friendly next.