Adding help modal
Because everybody needs a little help.

Well, this one is kinda large. But it is help!

There's also a bug fix in there for being able to add multiple
characters with the name 'Unknown'. Seemed important.

I'm pleased with where this is at. I could send it off to KM so
they could get a feel for the app and give me a yay or nay.

Maybe it's time to get back to my other hobbies...