Too many changes
Once again, I got distracted while working on a ....

SQUIRREL!!!

... Feature.

There is now an INFO button on each character that pops up a
Modal with some (currently un-styled) information. It's an
important feature for the future.

I've also changed the styling to have two columns. I think this
makes it a little more usable. However, it probably needs
some further styling to make it more clear that the two columns are
separated. A simple background color change would suffice.

So yeah, more style tweaks are required. The Rap Sheet should
probably be attached to the character popup too.