Adding weapons to profiles
Still needs better format in order to look reasonable...

But really, that's the whole site. Ugly.

Still, cool feature. At this point, you can add a character with
nearly all the information about them.

But that's just it. It's ONLY the information.

I don't play miniature games because of the information. I play
them because of the minis.