Toggle crew rules button is available
It doesn't work 100% yet. The primary issue I'm troubleshooting is
why if there is an invalid crew and you toggle to true, it gets
stuck at true. This appears to be a display component issue only.

Really, I need to be able to present the user with an Error
message.

... Because everybody loves errors explaining that something is
b0rk3d and they cannot fix it. #leSigh