Equipment integrated into the builder
Whew.

This thing is finally getting useful!

Unfortunately, it's only a display integration for the moment, but
we will soon be able to select a piece of equipment, and trigger
adding a piece of equipment to a character.

This is kinda a big deal.... Only kinda.

Another big part of the commit was refactoring more components
to be generic components. Others will follow suit shortly, but
I'm liking this approach a bunch.