First cut of assigning equipment functional
I cannot believe it. This is certainly too good to be true.

Many days of hacking now allow equipment to be assigned to characters.

Truly mind blowing. Like, your mind is a leaf infront of a freaking
jet engine and it is being blown into the next county, no, COUNTRY!

WHHEEEEEEEEEEEEEEEEEEW!!!!

Okay, testing isn't done, or even started, but wow. This feels
good!