Fixing funding and rep total
Well, I'm honestly cheating when calculating the rep and funding.

What should really happen is that there should be something done
after any action that triggers a re-calucation of the rep and
funding. That makes sense.

Instead, I'm currently updating it in two spots. That is bad.

Not sure how cross-cutting concerns like this fit into React/Redux.
That's something to research.