Required equipment now removed
This one was kinda simple, but complicated in other aspects.

Ultimately, I'm scared for my code's cleanliness. It is not bathing
frequently enough.

No, wait!

The real problem is that BMG has complicated crew building rules,
and it is complicated to implement them all.

However, I think this fixes the issue with removing equipment
when a character is removed without introducing side-effects.
So I am calling Issue #13 closed! Back to even with open bugs
and closed. Ooph...